# app/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: RDF Graphs Sharing/Exhange Server API Specifications
  description: "This is a RDF Graph Exhange Server REST API specifications to provide\
    \ users to access (CRUD - create, retrieve, update, and delete) RDF Graph-compliant\
    \ RDF resources.  - Each user's RDF graph (RDF Graph-compliant) can be shared\
    \ with other users to enable Federated SPARQL query if the aggregated user roles\
    \ and permissions over the RDF graphs passing the Access Control validation. Otherwise,\
    \ the access will be denied. - The create operation can accept data source formats\
    \ JSON-LD, RDF/XML, N3, or Turtle format for upload/create. - It is up to the\
    \ implementation to support multiple data source formats. - The RDF Graph Exchange\
    \ reference implementation will only support one of JSON-LD, N3, or Turtle formats."
  version: 0.1.3
servers:
- url: https://virtserver.swaggerhub.com/RS-API/RDF-Graph-Exchange/0.1.3
  description: SwaggerHub API Auto Mocking
paths:
...
  /api/graph/catalog:
    get:
      tags:
      - graph
      summary: Query the list of all the RDF graphs' names (URIs) and the response
        will be JSON format.
      operationId: catalog
      parameters:
      - name: access_token
        in: query
        description: Authorization access token string
        required: true
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/AccessToken'
      - name: userID
        in: query
        description: the ID of the organization of the client application
        required: true
        style: form
        explode: true
        schema:
          maxLength: 64
          minLength: 6
          type: string
          example: UCO-CASE-EU-CC-1234567890
          default: userUUID-unknown
      - name: query
        in: query
        description: Query GraphsSPARQL Query expression (max 1536). Note the common
          lowest limit for the entrie url is 2048 as the limit. The query SPARQL string
          must be url-encoded. The example below is not url-encoded to show the un-encoded
          SPARQL content.
        required: false
        style: form
        explode: true
        schema:
          maxLength: 256
          type: string
          nullable: true
          example: " SELECT ?g  WHERE { GRAPH ?g { ?s ?p ?o . } } "
          default: ""
      responses:
        "200":
          description: successful operation
          headers:
            x-next:
              description: A link to the next page of responses if the results is
                too large (per the Server implementation specifically).
              style: simple
              explode: false
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphListType'
        "400":
          description: "Bad request, exact description of error and any successful\
            \ results to be in content"
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "403":
          description: No permission to access object(s)
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "404":
          description: Specified object(s) not found
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "500":
          description: Server was not able to process request
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
      security: []
      x-openapi-router-controller: swagger_server.controllers.graph_controller
...
  /api/graph/create:
    post:
      tags:
      - graph
      summary: upload and create a new RDF graph upload RDF graph data source
      operationId: create_graph
      parameters:
      - name: access_token
        in: query
        description: Authorization access token string
        required: true
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/AccessToken'
      - name: userID
        in: query
        description: the ID of the organization of the client application
        required: true
        style: form
        explode: true
        schema:
          maxLength: 64
          minLength: 6
          type: string
          example: UCO-CASE-EU-CC-1234567890
          default: userUUID-unknown
      - name: graphURI
        in: query
        required: true
        style: form
        explode: true
        schema:
          maxLength: 128
          minLength: 6
          type: string
          example: <http://uco.example.org/bookStore>
          default: <http://uco.example.org/bookstore>
      - name: organizationID
        in: query
        description: the organization of the client application
        required: false
        style: form
        explode: true
        schema:
          maxLength: 64
          minLength: 6
          type: string
          example: uco-eu-cyber-123
          default: organizationID-unknown
      - name: rdfFormat
        in: query
        description: "the semantic graph format types, e.g., Turtle, RDF/XML, N-Triples,\
          \ JSON-LD, RDF/JSON, TriG, N-Quads, TriX. Note that parsing JSON-LD mostly\
          \ is much more time-consuming and expensive in computing than other formats.\
          \ This API supports other data formats, e.g., N3, RDF/XML, and TURTLE"
        required: true
        style: form
        explode: true
        schema:
          type: string
          nullable: false
          example: JSON-LD
          default: JSON-LD
          enum:
          - JSON-LD
          - TRIG
          - RDF/XML
          - N3
          - TURTLE
      - name: reference
        in: query
        description: the client applications' reference to this batch of RDF data
          to be uploaded and it is not used by the Server.
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: cs0123-b45678-2021-11-25-001
          default: n/a
      requestBody:
        description: upload RDF data source
        content:
          application/text:
            schema:
              $ref: '#/components/schemas/SourceDataType'
      responses:
        "200":
          description: successful operation
          content:
            application/text:
              schema:
                type: string
                description: Results in text format
                default: "{}"
                x-content-type: application/text
        "400":
          description: "Bad request, exact description of error and any successful\
            \ results to be in content"
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "403":
          description: No permission to access object(s)
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "404":
          description: Specified object(s) not found
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "500":
          description: Server was not able to process request
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
      security: []
      x-openapi-router-controller: swagger_server.controllers.graph_controller
  /api/graph/query:
    post:
      tags:
      - graph
      summary: Either SPARQL v1.1. RDF (Federated) Query or GraphQL (Federated) Query
        from multiple data sources
      operationId: query_graph
      parameters:
      - name: access_token
        in: query
        description: Authorization access token string
        required: true
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/AccessToken'
      - name: userID
        in: query
        description: the ID of the organization of the client application
        required: true
        style: form
        explode: true
        schema:
          maxLength: 64
          minLength: 6
          type: string
          example: UCO-CASE-EU-CC-1234567890
          default: userUUID-unknown
      - name: graphList
        in: query
        description: Either SPARQL or GraphQL query is supported
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/GraphListType'
      - name: organizationID
        in: query
        description: the organization of the client application
        required: false
        style: form
        explode: true
        schema:
          maxLength: 64
          minLength: 6
          type: string
          example: uco-eu-cyber-123
          default: organizationID-unknown
      - name: queryType
        in: query
        description: Either SPARQL or GraphQL query is supported
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: SPARQL
          enum:
          - SPARQL
          - GraphQL
      - name: reference
        in: query
        description: the client applications' reference to this batch of RDF data
          to be uploaded and it is not used by the Server.
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: cs0123-b45678-2021-11-25-001
          default: n/a
      requestBody:
        description: More specifics about what RDF data to delete using SPARQL 'DELETE
          DATA' statement. Note the Client can only delete the original Graph data
          sources being uploaded by the same User ID and Graph name. These graphs
          uploaded by others will not be deleted by the request as the protection
          by the ownership of those other graphs.
        content:
          application/text:
            schema:
              $ref: '#/components/schemas/SourceDataType'
      responses:
        "200":
          description: successful operation
          content:
            application/text:
              schema:
                type: string
                description: Results in text format
                default: "{}"
                x-content-type: application/text
        "400":
          description: "Bad request, exact description of error and any successful\
            \ results to be in content"
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "403":
          description: No permission to access object(s)
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "404":
          description: Specified object(s) not found
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "500":
          description: Server was not able to process request
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
      security: []
      x-openapi-router-controller: swagger_server.controllers.graph_controller
  /api/graph/update:
    post:
      tags:
      - graph
      summary: update RDF graph data source
      operationId: update_graph
      parameters:
      - name: access_token
        in: query
        description: Authorization access token string
        required: true
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/AccessToken'
      - name: userID
        in: query
        description: the ID of the organization of the client application
        required: true
        style: form
        explode: true
        schema:
          maxLength: 64
          minLength: 6
          type: string
          example: UCO-CASE-EU-CC-1234567890
          default: userUUID-unknown
      - name: graphURI
        in: query
        required: true
        style: form
        explode: true
        schema:
          maxLength: 128
          minLength: 6
          type: string
          example: <http://uco.example.org/bookStore>
          default: <http://uco.example.org/bookstore>
      - name: organizationID
        in: query
        description: the organization of the client application
        required: false
        style: form
        explode: true
        schema:
          maxLength: 64
          minLength: 6
          type: string
          example: uco-eu-cyber-123
          default: organizationID-unknown
      - name: reference
        in: query
        description: the client applications' reference to this batch of RDF data
          to be uploaded and it is not used by the Server.
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: cs0123-b45678-2021-11-25-001
          default: n/a
      requestBody:
        description: SPARQL v1.1 (see W3C SPARQL v1.1. Update Specification - https://www.w3.org/TR/sparql11-update/
          ) as the specification syntax of the payload SPARQL Update statements.
        content:
          application/text:
            schema:
              $ref: '#/components/schemas/SourceDataType'
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Results'
        "400":
          description: "Bad request, exact description of error and any successful\
            \ results to be in content"
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "403":
          description: No permission to access object(s)
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "404":
          description: Specified object(s) not found
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "500":
          description: Server was not able to process request
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
      security: []
      x-openapi-router-controller: swagger_server.controllers.graph_controller
  /api/graph/delete:
    post:
      tags:
      - graph
      summary: Delete RDF graph data source
      operationId: delete_graph
      parameters:
      - name: access_token
        in: query
        description: Authorization access token string
        required: true
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/AccessToken'
      - name: userID
        in: query
        description: the ID of the organization of the client application
        required: true
        style: form
        explode: true
        schema:
          maxLength: 64
          minLength: 6
          type: string
          example: UCO-CASE-EU-CC-1234567890
          default: userUUID-unknown
      - name: graphURI
        in: query
        required: true
        style: form
        explode: true
        schema:
          maxLength: 128
          minLength: 6
          type: string
          example: <http://uco.example.org/bookStore>
          default: <http://uco.example.org/bookstore>
      - name: organizationID
        in: query
        description: the organization of the client application
        required: false
        style: form
        explode: true
        schema:
          maxLength: 64
          minLength: 6
          type: string
          example: uco-eu-cyber-123
          default: organizationID-unknown
      - name: reference
        in: query
        description: the client applications' reference to this batch of RDF data
          to be uploaded and it is not used by the Server.
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: cs0123-b45678-2021-11-25-001
          default: n/a
      requestBody:
        description: More specifics about what RDF data to delete using SPARQL 'DELETE
          DATA' statement. Note the Client can only delete the original Graph data
          sources being uploaded by the same User ID and Graph name. These graphs
          uploaded by others will not be deleted by the request as the protection
          by the ownership of those other graphs.
        content:
          application/text:
            schema:
              $ref: '#/components/schemas/SourceDataType'
      responses:
        "200":
          description: successful operation
          content:
            application/text:
              schema:
                type: string
                description: Results in text format
                default: "{}"
                x-content-type: application/text
        "400":
          description: "Bad request, exact description of error and any successful\
            \ results to be in content"
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "403":
          description: No permission to access object(s)
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "404":
          description: Specified object(s) not found
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "500":
          description: Server was not able to process request
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
      security: []
      x-openapi-router-controller: swagger_server.controllers.graph_controller
...
components:
  schemas:
    AccessToken:
      type: string
      description: "Authorization Access Token in OAuth 2.0 and there are a few different\
        \ format. This schema assumes the existence of Authorization Access Server\
        \ such as OAuth 2.0 Servers besides the Server so that the Clients / Applications\
        \ / Users can request Access Token (short-lived or long-lived) to access this\
        \ the API services. And, the API Interface here will not prescribe the formats.\
        \ One of the common formats for Access Token is Json Web Token (JWT). A JWT\
        \ is represented as a sequence of base64url encoded values that are separated\
        \ by period characters with three parts as <Header>.<Payload>.<Signature>\
        \ - see the example value. And, the Interface will not prescribe the specific\
        \ format here."
      example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vdHJ1c3R5YXBwLmNvbS8iLCJleHAiOjEzMDA4MTkzODAsInN1YiI6InVzZXJzLzg5ODM0NjIiLCJzY29wZSI6InNlbGYgYXBpL2J1eSJ9.43DXvhrwMGeLLlP4P4izjgsBB2yrpo82oiUPhADakLs
...
    Errors:
      required:
      - code
      - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
...
    GraphListType:
      type: array
      description: Graph list
      example:
      - <http://uco.example.org/bookStore>
      - <http://uco.example.org/student>
      items:
        $ref: '#/components/schemas/GraphType'
...
    Results:
      type: string
      description: JSON as Query results
      example: "{}"
...
    SourceDataType:
      type: string
      example: " @prefix dc: <http://purl.org/dc/elements/1.1/> . @prefix ns: <http://example.org/ns#>\
        \ . <http://example/book2> ns:price 42 . <http://example/book2> dc:title \"\
        David Copperfield\" . <http://example/book2> dc:creator \"Edmund Wells\" ."
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
      x-basicInfoFunc: swagger_server.controllers.authorization_controller.check_BasicAuth
    BearerAuth:
      type: http
      scheme: bearer
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_BearerAuth
    ApiKeyAuth:
      type: apiKey
      name: X-API-Key
      in: header
      x-apikeyInfoFunc: swagger_server.controllers.authorization_controller.check_ApiKeyAuth
    OpenID:
      type: openIdConnect
      openIdConnectUrl: https://example.com/.well-known/openid-configuration
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://example.com/oauth/authorize
          tokenUrl: https://example.com/oauth/token
          scopes:
            read: Grants read access
            write: Grants write access
            admin: Grants access to admin operations
      x-tokenInfoFunc: swagger_server.controllers.authorization_controller.check_OAuth2
      x-scopeValidateFunc: swagger_server.controllers.authorization_controller.validate_scope_OAuth2